refactor(models,core): move model metadata + routing into domain folders
Move healthy single-domain modules into their final homes:
- src/models/ — metadata.ts, modelLimits.ts, modelCapabilities.ts, modelNames.ts
- src/core/ — routing.ts (transport types already live in core/transport.ts)

Update every importer (extension.ts, request/*, thinking/*, transports/*,
usage/*, tests, scripts/validate-models.ts, verify-estimate-token-count.ts).
Behavior-preserving; pure path churn.
