+++ src/test/modelLimits.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert/strict";
 import { describe, it } from "node:test";
-import { calculateModelLimits } from "../modelLimits.js";
+import { calculateModelLimits } from "../models/modelLimits.js";
 
 const metadata = {
   contextWindow: 100_000,
