+++ src/usage/tracker.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import type { ModelCost } from "../metadata";
+import type { ModelCost } from "../models/metadata";
 import type { TransportRequestSummary } from "../streaming";
 import { fetchGoUsage, mergeServerUsage, GO_USAGE_SYNC_TTL_MS, type GoUsageApiResponse } from "./goUsageSync";
 import {
